Extract numeric query option helpers
